modify: 헷갈리는 변수명 및 파일명 변경
